requirements.txt
flask==2.0.1
   requests==2.28.1
   pandas==1.4.3
   ta==0.10.2
   websocket-client==1.4.1
   gunicorn==20.1.0